Рефакторинг процесса запуска и конфигурирования приложений  Конфигурации для компонент описаны структурами  Добавлены тесты (первые, ура) для конфигураций  Параметры приложения, очередей и подключения к БД вынесены во флаги
