+++ cmd/clickhouse-sender/main.go
@@ -7,26 +7,46 @@ import (
 	"github.com/gofiber/fiber/v2"
 	"github.com/gofiber/fiber/v2/middleware/logger"
 	"github.com/gofiber/fiber/v2/middleware/monitor"
+	"github.com/gofiber/fiber/v2/middleware/pprof"
 	"github.com/gofiber/fiber/v2/middleware/requestid"
-	"github.com/vsurkov/go-metr/internal/buffer"
-	"github.com/vsurkov/go-metr/internal/db"
-	"github.com/vsurkov/go-metr/internal/instance"
-	"github.com/vsurkov/go-metr/internal/rabbitmq"
+	"github.com/vsurkov/go-metr/internal/app/instance"
+	"github.com/vsurkov/go-metr/internal/common/buffer"
+	"github.com/vsurkov/go-metr/internal/platform/database"
+	rabbitmq2 "github.com/vsurkov/go-metr/internal/platform/rabbitmq"
 	"log"
 	"time"
 )
 
-var app = new(instance.Instance)
+var app instance.Instance
 
 var (
-	app_port     = flag.String("app_port", "4000", "Application port, default http://127.0.0.1:4000")
-	uri          = flag.String("uri", "amqp://rabbitmq:rabbitmq@localhost:5672/", "AMQP URI")
-	exchange     = flag.String("exchange", "test-exchange", "Durable, non-auto-deleted AMQP exchange name")
-	exchangeType = flag.String("exchange-type", "direct", "Exchange type - direct|fanout|topic|x-custom")
-	queue        = flag.String("queue", "rncb.events.queue", "Ephemeral AMQP queue name")
-	bindingKey   = flag.String("key", "test-key", "AMQP binding key")
-	consumerTag  = flag.String("consumer-tag", "simple-consumer", "AMQP consumer tag (should not be blank)")
-	bufferSize   = flag.Int("bufferSize", 100, "Buffer size for batch database Writing (default 100 messages)")
+	app_port              = flag.Int("app_port", 4000, "Application port, default 4000")
+	app_full_name         = flag.String("app_full_name", "rest-application", "Application name, default 'rest-application'")
+	app_name              = flag.String("app_name", "rncb", "Application name, will be used in Queue prefix, and DB name default 'rncb'")
+	app_server_logging    = flag.Bool("app_server_logging", false, "Enabling Fiber http server requests logging, default 'false'")
+	app_enable_profiling  = flag.Bool("abb_enable_profiling", false, "Enabling runtime profiling data in the format expected by the pprof visualization tool on http://host:port/debug/pprof/profile, default 'false'")
+	app_enable_request_id = flag.Bool("", false, "Enabling server middleware that adds an identifier to the response, default 'false'")
+
+	rabbit_host         = flag.String("rabbit_host", "localhost", "IP address or hostname for connecting to the RabbitMQ, default 'localhost'")
+	rabbit_port         = flag.Int("rabbit_port", 5672, "IP address or hostname for connecting to the RabbitMQ, default '5672'")
+	rabbit_user         = flag.String("rabbit_user", "rabbitmq", "User name for connecting to the RabbitMQ, default 'rabbitmq'")
+	rabbit_password     = flag.String("rabbit_password", "rabbitmq", "Password for connecting to the RabbitMQ, default 'rabbitmq'")
+	rabbit_exchange     = flag.String("rabbit_exchange", "exchange", "Durable, non-auto-deleted AMQP exchange name, default 'exchange'")
+	rabbit_exchangeType = flag.String("rabbit_exchangeType", "direct", "Exchange type - direct|fanout|topic|x-custom, default direct")
+	rabbit_bindingKey   = flag.String("rabbit_bindingKey", "key", "AMQP binding key")
+	rabbit_consumerTag  = flag.String("rabbit_consumerTag", "simple-consumer", "AMQP consumer tag (should not be blank)")
+
+	db_host              = flag.String("db_host", "localhost", "IP address or hostname for connecting to the Clickhouse, default 'localhost'")
+	db_port              = flag.Int("db_port", 9000, "IP address or hostname for connecting to the Clickhouse, default '9000'")
+	db_default_database  = flag.String("db_default_database", "default", "Default database name into Clickhouse, default 'default'")
+	db_user              = flag.String("db_user", "default", "User name for connecting to the Clickhouse, default 'default'")
+	db_password          = flag.String("db_password", "", "Password for connecting to the Clickhouse, default ''")
+	db_debug             = flag.Bool("db_debug", false, "Enable debug mode for database, default false")
+	db_dial_timeout      = flag.Duration("db_dial_timeout", time.Second, "Dial database timeout, default 1 second")
+	db_max_open_conns    = flag.Int("db_max_open_conns", 10, "Maximum open connections to database, default db_max_idle_conns +5")
+	db_max_idle_conns    = flag.Int("db_max_idle_conns", 5, "Maximum idle connections to database, default 5")
+	db_conn_max_lifetime = flag.Duration("db_conn_max_lifetime", time.Hour, "Maximum lifetime for connections to database, default 1 hour")
+	bufferSize           = flag.Int("bufferSize", 100, "Buffer size for batch database Writing (default 100 messages)")
 )
 
 func init() {
@@ -34,21 +54,31 @@ func init() {
 }
 
 func main() {
-	app.Version = "0.0.1"
-	app.Name = "rncb"
+
+	app = instance.Instance{
+		Config: instance.Config{
+			Port:     *app_port,
+			Name:     *app_name,
+			FullName: *app_full_name,
+			Version:  "0.0.1",
+		},
+		DB:           database.Database{},
+		RB:           rabbitmq2.Rabbit{},
+		KnownSystems: nil,
+	}
 
 	// Clickhouse configuration
-	clickhouseDB, err := new(db.Database).Connect(db.DBConfig{
-		Host:              "127.0.0.1",
-		Port:              9000,
-		Database:          "default",
-		Username:          "default",
-		Password:          "",
-		Debug:             false,
-		DialTimeout:       time.Second,
-		MaxOpenConns:      10,
-		MaxIdleConns:      5,
-		ConnMaxLifetime:   time.Hour,
+	clickhouseDB, err := new(database.Database).NewConnection(database.Config{
+		Host:              *db_host,
+		Port:              *db_port,
+		Database:          *db_default_database,
+		User:              *db_user,
+		Password:          *db_password,
+		Debug:             *db_debug,
+		DialTimeout:       *db_dial_timeout,
+		MaxOpenConns:      *db_max_open_conns,
+		MaxIdleConns:      *db_max_idle_conns,
+		ConnMaxLifetime:   *db_conn_max_lifetime,
 		CompressionMethod: clickhouse.CompressionLZ4,
 	})
 	if err != nil {
@@ -59,16 +89,20 @@ func main() {
 	app.DB.Buffer = b.NewBuffer(*bufferSize)
 
 	// RabbitMQ configuration
-	app.RB.Cfg = &rabbitmq.Config{
-		Uri:          *uri,
-		Exchange:     *exchange,
-		ExchangeType: *exchangeType,
-		BindingKey:   *bindingKey,
-		ConsumerTag:  *consumerTag,
+	app.RB.Config = &rabbitmq2.Config{
+		Host:         *rabbit_host,
+		Port:         *rabbit_port,
+		User:         *rabbit_user,
+		Password:     *rabbit_password,
+		Exchange:     *rabbit_exchange,
+		ExchangeType: *rabbit_exchangeType,
+		Queue:        fmt.Sprintf("%v.events.queue", app.Config.Name),
+		BindingKey:   *rabbit_bindingKey,
+		ConsumerTag:  *rabbit_consumerTag,
 	}
 
 	//	TODO сделать множественные очереди
-	cons, err := rabbitmq.NewConsumer(app.RB.Cfg, *queue, app.DB)
+	cons, err := rabbitmq2.NewConsumer(app.RB.Config, app.RB.Config.Queue, app.DB)
 	if err != nil {
 		log.Fatalf("%s", err)
 	}
@@ -84,12 +118,19 @@ func main() {
 	// Fiber configuration
 
 	a := fiber.New(fiber.Config{
-		AppName: fmt.Sprintf("go-metr v %v, app %v", app.Version, app.Name),
+		AppName: fmt.Sprintf("go-metr v %v, app %v", app.Config.Version, app.Config.FullName),
 	})
 
 	// Fiber middleware configuration
-	a.Use(logger.New())
-	a.Use(requestid.New())
+	if *app_server_logging {
+		a.Use(logger.New())
+	}
+	if *app_enable_request_id {
+		a.Use(requestid.New())
+	}
+	if *app_enable_profiling {
+		a.Use(pprof.New())
+	}
 
 	// Fiber endpoints configuration
 	a.Get("/", func(ctx *fiber.Ctx) error {
