+++ go.mod
@@ -6,21 +6,25 @@ require (
 	github.com/ClickHouse/clickhouse-go/v2 v2.2.0
 	github.com/gofiber/fiber/v2 v2.36.0
 	github.com/google/uuid v1.3.0
-	github.com/mailru/easyjson v0.7.7
 	github.com/streadway/amqp v1.0.0
+	github.com/stretchr/testify v1.7.5
 )
 
 require (
 	github.com/andybalholm/brotli v1.0.4 // indirect
-	github.com/josharian/intern v1.0.0 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/klauspost/compress v1.15.0 // indirect
+	github.com/kr/pretty v0.3.0 // indirect
 	github.com/paulmach/orb v0.7.1 // indirect
 	github.com/pierrec/lz4/v4 v4.1.15 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/shopspring/decimal v1.3.1 // indirect
 	github.com/valyala/bytebufferpool v1.0.0 // indirect
 	github.com/valyala/fasthttp v1.39.0 // indirect
 	github.com/valyala/tcplisten v1.0.0 // indirect
 	go.opentelemetry.io/otel v1.9.0 // indirect
 	go.opentelemetry.io/otel/trace v1.9.0 // indirect
-	golang.org/x/sys v0.0.0-20220429233432-b5fbb4746d32 // indirect
+	golang.org/x/sys v0.0.0-20220520151302-bc2c85ada10a // indirect
+	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
